refactor(update2.py): refactor out printing functionality
to facilitate addition of other commands
