+++ src/btrfs2s3/_internal/commands/update2.py
@@ -19,7 +19,6 @@
 
 from __future__ import annotations
 
-from collections import defaultdict
 from contextlib import ExitStack
 from enum import auto
 from enum import Enum
@@ -29,30 +28,16 @@ from typing import cast
 from typing import TYPE_CHECKING
 from zoneinfo import ZoneInfo
 
-import arrow
 from boto3.session import Session
-from rich.box import HORIZONTALS
-from rich.columns import Columns
-from rich.console import Console
-from rich.console import Group
-from rich.filesize import decimal
-from rich.highlighter import ISO8601Highlighter
-from rich.markup import escape
 from rich.prompt import Confirm
 from rich.prompt import Prompt
-from rich.table import Column
-from rich.table import Table
-from rich.text import Text
-from rich.tree import Tree
 
+from btrfs2s3._internal.commands.printing import print_plan
 from btrfs2s3._internal.config import load_from_path
-from btrfs2s3._internal.cvar import TZINFO
 from btrfs2s3._internal.cvar import use_tzinfo
 from btrfs2s3._internal.piper import filter_pipe
 from btrfs2s3._internal.planner import Actions
 from btrfs2s3._internal.planner import assess
-from btrfs2s3._internal.planner import AssessedBackup
-from btrfs2s3._internal.planner import AssessedSnapshot
 from btrfs2s3._internal.planner import Assessment
 from btrfs2s3._internal.planner import assessment_to_actions
 from btrfs2s3._internal.planner import ConfigTuple
@@ -62,252 +47,16 @@ from btrfs2s3._internal.planner import SnapshotDir
 from btrfs2s3._internal.planner import Source
 from btrfs2s3._internal.preservation import Params
 from btrfs2s3._internal.preservation import Policy
-from btrfs2s3._internal.preservation import TS
 from btrfs2s3._internal.resolver import Flags
-from btrfs2s3._internal.resolver import KeepMeta
-from btrfs2s3._internal.resolver import Reasons
-from btrfs2s3._internal.time_span_describer import describe_time_span
 
 if TYPE_CHECKING:
     import argparse
-    from collections.abc import Collection
-    from collections.abc import Iterable
     from collections.abc import Iterator
-    from collections.abc import Sequence
     from typing import TypedDict
 
+    from rich.console import Console
+
     from btrfs2s3._internal.config import Config
-
-
-_iso8601_highlight = ISO8601Highlighter()
-
-
-def _time_span_key(time_span: TS) -> tuple[float, float]:
-    start, end = time_span
-    return (start - end, start)
-
-
-def _describe_time_spans(time_spans: Iterable[TS]) -> Text:
-    return describe_time_span(sorted(time_spans, key=_time_span_key)[0], bounds="[]")
-
-
-def _describe_preserve(keep_meta: KeepMeta) -> Text:
-    if keep_meta.reasons & Reasons.Preserved:
-        return _describe_time_spans(keep_meta.time_spans)
-    if keep_meta.reasons & Reasons.MostRecent:
-        return Text("<most recent>", style="keep")
-    if keep_meta.reasons & Reasons.SendAncestor:
-        return Text("<ancestor>", style="keep")
-    if keep_meta.reasons:  # pragma: no cover
-        return Text("<keep!>", style="keep")
-    return Text("<not keeping>", style="not_keeping")
-
-
-def _describe_time(time: float) -> Text:
-    return _iso8601_highlight(
-        arrow.get(time, tzinfo=TZINFO.get()).format("YYYY-MM-DDTHH:mm:ss")
-    )
-
-
-def _keep_emoji(keep_meta: KeepMeta) -> Text:
-    if keep_meta.flags & Flags.New:
-        return Text.from_markup(":sparkles:")
-    if not keep_meta.reasons:
-        return Text.from_markup(":skull:")
-    return Text()
-
-
-def _make_snapshots_table(snapshots: Collection[AssessedSnapshot]) -> Table:
-    table = Table(
-        Column("path"),
-        Column("ctime"),
-        Column("ctransid", style="ctransid"),
-        Column("preserve"),
-        title="snapshots",
-        title_justify="left",
-        row_styles=["none", "dim"],
-        box=HORIZONTALS,
-    )
-    snapshots = sorted(
-        snapshots, key=lambda k: (k.snapshot_dir.path, k.source.path, k.info.ctransid)
-    )
-    for snap in snapshots:
-        path = (
-            Text(style="" if snap.meta.reasons else "delete")
-            .append(_keep_emoji(snap.meta))
-            .append(escape(str(snap.snapshot_dir.get_path(snap.info.id))))
-        )
-        ctime = _describe_time(snap.info.ctime)
-        ctransid = str(snap.info.ctransid)
-        preserve = _describe_preserve(snap.meta)
-        table.add_row(path, ctime, ctransid, preserve)
-    return table
-
-
-def _make_new_snapshots_table(asmt: Assessment) -> Table | None:
-    new_snaps = sorted(
-        (s for s in asmt.snapshots.values() if s.meta.flags & Flags.New),
-        key=lambda s: (s.source.path, s.snapshot_dir.get_path(s.info.id)),
-    )
-    if not new_snaps:
-        return None
-    table = Table(
-        Column("source"),
-        Column("path"),
-        title=":sparkles:newly-created snapshots:sparkles:",
-        title_justify="left",
-        row_styles=["none", "dim"],
-        box=HORIZONTALS,
-    )
-    for new_snap in new_snaps:
-        source = escape(str(new_snap.source.path))
-        path = escape(str(new_snap.snapshot_dir.get_path(new_snap.info.id)))
-        table.add_row(source, path)
-    return table
-
-
-def _backup_key(backup: AssessedBackup) -> tuple[Path, int, float]:
-    return (backup.source.path, backup.info.ctransid, backup.info.ctime)
-
-
-def _make_backups_tree(backups: Collection[AssessedBackup]) -> Tree:
-    uuid_to_children: dict[tuple[Remote, bytes], list[AssessedBackup]] = defaultdict(
-        list
-    )
-    remote_full_backups: dict[Remote, list[AssessedBackup]] = defaultdict(list)
-    for backup in backups:
-        send_parent_uuid = backup.info.send_parent_uuid
-        if send_parent_uuid:
-            uuid_to_children[(backup.remote, send_parent_uuid)].append(backup)
-        else:
-            remote_full_backups[backup.remote].append(backup)
-
-    def add_backup_node(parent: Tree, backup: AssessedBackup) -> None:
-        key = Text(no_wrap=True, overflow="ellipsis").append(
-            _keep_emoji(backup.meta).append(
-                escape(backup.key), style="key" if backup.meta.reasons else "delete"
-            )
-        )
-        stats = [
-            _describe_preserve(backup.meta),
-            _describe_time(backup.info.ctime),
-            Text(str(backup.info.ctransid), style="ctransid"),
-        ]
-        size = backup.stat and backup.stat.size
-        if size is not None:
-            stats.append(Text(decimal(size), style="cost"))
-        storage_class = backup.stat and backup.stat.storage_class
-        if storage_class:
-            stats.append(Text(storage_class, style="cost"))
-        info = Columns(stats)
-        node = parent.add(Group(key, info))
-        children = uuid_to_children.get((backup.remote, backup.info.uuid), ())
-        for child in sorted(children, key=_backup_key):
-            add_backup_node(node, child)
-
-    tree = Tree("backups")
-    for remote in sorted(remote_full_backups.keys(), key=lambda r: r.name):
-        remote_tree = tree.add(
-            Text.from_markup(":cloud-emoji:").append(escape(remote.name))
-        )
-        for backup in sorted(remote_full_backups[remote], key=_backup_key):
-            add_backup_node(remote_tree, backup)
-    return tree
-
-
-def _make_action_tables(actions: Actions) -> Sequence[Table]:
-    rename_snapshots = Table(
-        Column("action"),
-        Column("source"),
-        Column("arrow"),
-        Column("dest"),
-        box=None,
-        show_header=False,
-        row_styles=["none", "dim"],
-    )
-    for rename_args in actions.rename_snapshots:
-        rename_snapshots.add_row(
-            Text.from_markup(":pencil-emoji: rename:"),
-            escape(str(rename_args.snapshot_dir.get_path(rename_args.info.id))),
-            "->",
-            escape(rename_args.target_name),
-            style="modify",
-        )
-
-    upload_backups = Table(
-        Column("action"),
-        Column("source"),
-        Column("arrow"),
-        Column("remote"),
-        Column("key"),
-        box=None,
-        show_header=False,
-        row_styles=["none", "dim"],
-    )
-    for upload_args in actions.upload_backups:
-        upload_backups.add_row(
-            Text.from_markup(":cloud-emoji: upload:"),
-            escape(str(upload_args.snapshot_dir.get_path(upload_args.info.id))),
-            "->",
-            Text.from_markup(":cloud-emoji:").append(escape(upload_args.remote.name)),
-            escape(upload_args.key),
-            style="create",
-        )
-
-    destroy_snapshots = Table(
-        Column("action"),
-        Column("path"),
-        box=None,
-        show_header=False,
-        row_styles=["none", "dim"],
-    )
-    for destroy_args in actions.destroy_snapshots:
-        destroy_snapshots.add_row(
-            Text.from_markup(":skull: destroy:"),
-            escape(str(destroy_args.snapshot_dir.get_path(destroy_args.info.id))),
-            style="delete",
-        )
-
-    delete_backups = Table(
-        Column("action"),
-        Column("remote"),
-        Column("key"),
-        box=None,
-        show_header=False,
-        row_styles=["none", "dim"],
-    )
-    for delete_args in actions.delete_backups:
-        delete_backups.add_row(
-            Text.from_markup(":skull: delete:"),
-            Text.from_markup(":cloud-emoji:").append(escape(delete_args.remote.name)),
-            escape(delete_args.key),
-            style="delete",
-        )
-
-    return (rename_snapshots, upload_backups, destroy_snapshots, delete_backups)
-
-
-def print_plan(*, console: Console, assessment: Assessment, actions: Actions) -> None:
-    console.print(_make_snapshots_table(assessment.snapshots.values()))
-    console.print()
-    console.print(_make_backups_tree(assessment.backups.values()))
-    console.print()
-
-    new_snapshots_table = _make_new_snapshots_table(assessment)
-    if new_snapshots_table:
-        console.print(new_snapshots_table)
-        console.print()
-
-    action_tables = _make_action_tables(actions)
-    if any(t.row_count for t in action_tables):
-        console.print("actions to take:")
-        for table in action_tables:
-            if table.row_count:
-                console.print(table)
-        console.print()
-    else:
-        console.print("nothing to be done!")
-        console.print()
 
 
 NAME = "update"
